refactor(engine): share table-function call parsing
