+++ crates/coral-engine/src/runtime/mod.rs
@@ -11,4 +11,5 @@ pub(crate) mod query;
 pub(crate) mod query_planner;
 pub(crate) mod registry;
 pub(crate) mod schema_provider;
+pub(crate) mod scoped_table_functions;
 pub(crate) mod source_functions;
